Validate JAX port against MATLAB source and fix discrepancies
Line-by-line comparison against the MATLAB reference
(bayesian-flame-impulse-response) surfaced several correctness bugs and
missing pieces in the JAX port; this fixes all of them:

- core/cost.py: fix q[valid] -> q[valid:] truncation bug that broke the
  MAP cost/gradient/Hessian computation, and restructure
  calculate_cost/calculate_cost_varpro into thin dispatchers + jitted
  cores so `data_level` and `valid` never enter the jit trace (a single
  static_argnums can't mark only some leaves of a dict argument static).
  Also drops a dead throwaway convolution and jnp.clip gamma clamps that
  had no MATLAB counterpart.
- signals/prepare.py: replace the bandwidth/PSD-based auto-downsampling
  (ported from a block MATLAB explicitly disabled as "legacy v1.0") with
  the actual active DSmode/DSvalue switch from prepareSignals.m.
- config/defaults.py: was an empty stub; add PreprocConfig and reconcile
  PriorConfig/OptimizerConfig defaults (LFL_sigma, optimizer tolerances,
  finalHessianMethod) with loadDefaultConfig.m. compute_final_hessian
  now implements the real GN/BFGS/FD switch instead of always GN(+BFGS).
- inference/optimizer.py: add the ill-conditioned-Hessian bailout
  MATLAB's LM loop has and the port had dropped.
- core/ftf.py: new, ports calculateFTF.m (gain/phase + Monte Carlo
  credible intervals), previously missing entirely.
- Enable jax_enable_x64 and fix stray float32 casts throughout, to match
  MATLAB's double precision.
- utils/plotting.py, utils/tfdsi.py: were empty stubs; add matplotlib
  ports of getColours/errorpatch/cornerHeatmap/plotModelComparisonH and
  an approximate, documented substitute for the MATLAB System ID
  Toolbox-based TFDSI.m baseline.

Also fixes two bugs the fixes above surfaced via smoke testing:
- core/impulse_response.py, core/cost.py used bare intra-package imports
  that only work with subdirectories individually on sys.path,
  inconsistent with the __init__.py files' fully-qualified imports; the
  package could not be imported as committed.
- PriorConfig was a plain mutable dataclass and therefore unhashable,
  breaking every jit(static_argnums=...) call that included it.

Verified end-to-end with synthetic data: signal prep, prior generation,
both cost functions, the LM optimizer in serial and vmapped modes, all
three final_hessian_method branches, FTF, TFDSI, and plotting.

+++ inference/mcmc.py
@@ -339,7 +339,7 @@ def run_mcmc_python(log_post_fn : Callable,
     L     = np.linalg.cholesky(np.array(prop_cov))
 
     # Pre-allocate in NumPy for speed
-    chain        = np.zeros((d, n_iter), dtype=np.float32)
+    chain        = np.zeros((d, n_iter), dtype=np.float64)
     log_post_arr = np.full(n_iter, -np.inf, dtype=np.float64)
     accepted_arr = np.zeros(n_iter, dtype=bool)
 
@@ -353,7 +353,7 @@ def run_mcmc_python(log_post_fn : Callable,
         # Propose
         z       = rng.standard_normal(d)
         b_prop  = b_cur + L @ z
-        lp_prop = float(log_post_fn(jnp.array(b_prop, dtype=jnp.float32)))
+        lp_prop = float(log_post_fn(jnp.array(b_prop, dtype=jnp.float64)))
 
         if not np.isfinite(lp_prop):
             lp_prop = -np.inf
